refactor(fe): changed some naming back to previous
