+++ client/src/components/Dashboard/DashboardIllustration/DashboardIllustration.module.scss
